CHORE: re declare friend event dto
