+++ kafka/src/main/kotlin/vp/togedo/kafka/data/friend/FriendRequestEventDto.kt
@@ -1,0 +1,12 @@
+package vp.togedo.kafka.data.friend
+
+import com.fasterxml.jackson.annotation.JsonProperty
+
+data class FriendRequestEventDto(
+    @JsonProperty("receiverId")
+    val receiverId: String,
+    @JsonProperty("sender")
+    val sender: String,
+    @JsonProperty("image")
+    val image: String?
+)
